+++ code/model/decoding.cpp
@@ -3,41 +3,20 @@
 
 using namespace std;
 
-void print_model_params(){
-    cout << "Pis:" << endl;
-    for(int i = 0; i < 3; i++){
-        cout << exp(pis[i]) << " ";
-    }
-
-    cout << endl << "Trans:" << endl;
-    for(int i = 0; i < 3; i++){
-        for(int j = 0; j < 3; j++){
-            cout << exp(trans[i][j]) << " ";
-        }
-        cout << endl;
-    }
-
-    cout << "emission_M" << endl;
-    for(int i = 0; i < 4; i++){
-        for(int j = 0; j < 4; j++){
-            cout << exp(emission_M[i][j]) << " ";
-        }
-        cout << endl;
-    }
-
-    cout << "emission_Ix" << endl;
-    for(int j = 0; j < 4; j++){
-        cout << exp(emission_Ix[j]) << " ";
-    }
-    cout << endl;
-
-    cout << "emission_Iy" << endl;
-    for(int j = 0; j < 4; j++){
-        cout << exp(emission_Ix[j]) << " ";
-    }
-    cout << endl;
-}
-
+/**
+ * @brief Function for aligning sequences using the Viterbi algorithm.
+ * 
+ * This function takes a pair of sequences and aligns them using the Viterbi algorithm with 
+ * the pre-trained model parameters. The aligned sequences are then saved to the specified 
+ * file path.
+ * 
+ * @param pair The pair of strings using the std::pair class from the C++ Standard Library.
+ * @param file_path The file path where the aligned sequences will be saved.
+ * 
+ * @note The model parameters must be provided beforehand.
+ * 
+ * @author Leon Hegedić
+ */
 void viterbi(pair<string, string> pair, const string& file_path){
     string x = pair.first;
     string y = pair.second;
@@ -57,14 +36,14 @@ void viterbi(pair<string, string> pair, const string& file_path){
 
     for(int i = 1; i < n + 1; i++){
         log_M[i][0] = -numeric_limits<double>::infinity();
-        log_X[i][0] = -numeric_limits<double>::infinity();//i * epsilon + delta;
+        log_X[i][0] = -numeric_limits<double>::infinity();
         log_Y[i][0] = -numeric_limits<double>::infinity();
     }
 
     for(int j = 1; j < m + 1; j++){
         log_M[0][j] = -numeric_limits<double>::infinity();
         log_X[0][j] = -numeric_limits<double>::infinity();
-        log_Y[0][j] = -numeric_limits<double>::infinity();//j * epsilon + delta;
+        log_Y[0][j] = -numeric_limits<double>::infinity();
     }
 
     double v1, v2, v3, v4, v5, v6, v7;
@@ -127,6 +106,6 @@ void viterbi(pair<string, string> pair, const string& file_path){
         }
     }
 
-    cout << "Viterbi score is: " << score << endl;
+    cout << "Viterbi score: " << score << endl;
     save_pair_into_file(file_path, make_pair(align_x, align_y));
 }
