+++ model_ensure_safety
@@ -1,0 +1,36 @@
+1
+0
+0
+0.996837
+0.00126502
+0.00181847
+0.0969697
+0.90303
+0
+0.129944
+0
+0.870056
+0.292615
+0.0115433
+0.0313093
+0.00948767
+0.013599
+0.137808
+0.00553447
+0.0158918
+0.0313884
+0.00561354
+0.194181
+0.00672043
+0.00925047
+0.0190544
+0.00608792
+0.209915
+0.393939
+0.169697
+0.169697
+0.266667
+0.384181
+0.186441
+0.248588
+0.180791
